게시글 댓글 보유 시 삭제 안되게 변경 및 board/all html 추가

+++ src/main/java/org/maengle/board/services/BoardDeleteService.java
@@ -7,6 +7,8 @@ import org.maengle.file.services.FileDeleteService;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDateTime;
+
 @Lazy
 @Service
 @RequiredArgsConstructor
@@ -17,7 +19,14 @@ public class BoardDeleteService {
 
     public void process(Long seq) {
         BoardData item = infoService.get(seq);
+        if (item.getCommentCount() > 0) { // 댓글이 남아 있는 경우 소프트 삭제
+            item.setDeletedAt(LocalDateTime.now());
+            repository.saveAndFlush(item);
 
+            return;
+        }
+
+        // 댓글이 남아있지 않은 경우는 DB에서 완전 분리
         String gid = item.getGid();
         repository.delete(item);
         deleteService.process(gid);
